# 8264288: Performance issue with MethodHandle.asCollector