--- a/src/java.base/share/classes/java/lang/invoke/LambdaFormEditor.java
+++ b/src/java.base/share/classes/java/lang/invoke/LambdaFormEditor.java
@@ -648,57 +648,6 @@ class LambdaFormEditor {
         return putInCache(key, form);
     }
 
-    LambdaForm collectArgumentArrayForm(int pos, MethodHandle arrayCollector) {
-        MethodType collectorType = arrayCollector.type();
-        int collectorArity = collectorType.parameterCount();
-        assert(arrayCollector.intrinsicName() == Intrinsic.NEW_ARRAY);
-        Class<?> arrayType = collectorType.returnType();
-        Class<?> elementType = arrayType.getComponentType();
-        BasicType argType = basicType(elementType);
-        int argTypeKey = argType.ordinal();
-        if (argType.basicTypeClass() != elementType) {
-            // return null if it requires more metadata (like String[].class)
-            if (!elementType.isPrimitive())
-                return null;
-            argTypeKey = TYPE_LIMIT + Wrapper.forPrimitiveType(elementType).ordinal();
-        }
-        assert(collectorType.parameterList().equals(Collections.nCopies(collectorArity, elementType)));
-        byte kind = COLLECT_ARGS_TO_ARRAY;
-        TransformKey key = TransformKey.of(kind, pos, collectorArity, argTypeKey);
-        LambdaForm form = getInCache(key);
-        if (form != null) {
-            assert(form.arity == lambdaForm.arity - 1 + collectorArity);
-            return form;
-        }
-        LambdaFormBuffer buf = buffer();
-        buf.startEdit();
-
-        assert(pos + 1 <= lambdaForm.arity);
-        assert(pos > 0);  // cannot filter the MH arg itself
-
-        Name[] newParams = new Name[collectorArity];
-        for (int i = 0; i < collectorArity; i++) {
-            newParams[i] = new Name(pos + i, argType);
-        }
-        Name callCombiner = new Name(new NamedFunction(arrayCollector, Intrinsic.NEW_ARRAY),
-                                        (Object[]) /*...*/ newParams);
-
-        // insert the new expression
-        int exprPos = lambdaForm.arity();
-        buf.insertExpression(exprPos, callCombiner);
-
-        // insert new arguments
-        int argPos = pos + 1;  // skip result parameter
-        for (Name newParam : newParams) {
-            buf.insertParameter(argPos++, newParam);
-        }
-        assert(buf.lastIndexOf(callCombiner) == exprPos+newParams.length);
-        buf.replaceParameterByCopy(pos, exprPos+newParams.length);
-
-        form = buf.endEdit();
-        return putInCache(key, form);
-    }
-
     LambdaForm filterArgumentForm(int pos, BasicType newType) {
         TransformKey key = TransformKey.of(FILTER_ARG, pos, newType.ordinal());
         LambdaForm form = getInCache(key);