… test/jdk/java/lang/invoke/MethodHandlesArityLimitsTest.java
… test/jdk/java/lang/invoke/VarargsArrayTest.java
… test/micro/org/openjdk/bench/java/lang/invoke/TypedAsCollector.java