@JornVernee 2021-04-05
Addressed latest review comments:
- Reverted back to using an injected constructor handle
- Added lambda form sharing
- Added a test case for a custom class to VarargsArrayTest, which catches the illegal access error in the intrinsified case (though the itrinsic itself is now removed).

I also had to switch back to the un-cached version for creating the array element setter, since the cached version casts the array to be a specific type, and if the lambda form is shared, this won't work (e.g. casting an Object[] to String[], depending on the order of creating the collectors). Adding caching there is left as a followup.